+++ src/app/api/articles/[slug]/route.ts
@@ -1,6 +1,6 @@
 import { createClient } from '@/utils/supabase/server';
 import { NextResponse } from 'next/server';
-import { categories } from '../route';
+import { CATEGORIES } from '@/app/components/Categories';
 
 // GET single article by slug (public)
 export async function GET(
