+++ src/app/components/Categories.tsx
@@ -1,0 +1,43 @@
+export const CATEGORIES = {
+    Destinations: [
+        "Cebu Highlights",
+        "Beaches & Islands",
+        "Mountain Escapes",
+        "Heritage & History",
+        "Hidden Gems",
+        "Travel Itineraries",
+    ],
+    "Brands and Products": [
+        "Homegrown Brands",
+        "Fashion & Apparel",
+        "Tech & Gadgets",
+        "Beauty & Wellness",
+        "Food Products",
+        "Eco-Friendly & Sustainable",
+    ],
+    Stories: [
+        "Life in Cebu",
+        "Resilience & Recovery",
+        "Student Stories",
+        "Entrepreneur Journeys",
+        "Cultural Narratives",
+        "Inspirational Profiles",
+    ],
+    "News and Entertainment": [
+        "Breaking News Cebu",
+        "Local Governance",
+        "Festivals & Events",
+        "Entertainment Buzz",
+        "Music & Arts",
+        "Sports",
+        "Campus News",
+    ],
+    Food: [
+        "Cebu Favorites",
+        "Street Food Finds",
+        "Café & Coffee Spots",
+        "Seafood Specials",
+        "Sweet Treats & Desserts",
+        "Food Reviews",
+    ],
+};
